fix(Controller): 동일 컨트롤러 사용으로 인한 오류 해결
